App/requirements.txt
keras==2.11.0
livelossplot==0.5.5
numpy==1.21.6
opencv_python==4.6.0.66
Pillow==9.3.0
streamlit==1.15.1
tensorflow==2.11.0
